+++ tests/photons_socket_tests/test_fake.py
@@ -26,7 +26,7 @@ describe AsyncTestCase, "Memory target":
 
         final_future = asyncio.Future()
         everything = {
-              "final_future": lambda: final_future
+              "final_future": final_future
             , "protocol_register": protocol_register
             }
         meta = Meta(everything, []).at("target")